+++ CHANGELOG.md
@@ -52,7 +52,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    - `Indices` and `Chunks` implement `IntoParallelIter`
    - **Breaking**: array subset iterators are moved into public `array_subset::iterators` and no longer in the `array_subset` namespace
  - Add a fast path to `Array::retrieve_chunk_subset{_opt}` if the entire chunk is requested
- - `DimensionName::new()` now accepts anything implementing `Into<String>`
+ - `DimensionName::new()` generalised to accept a name implementing `Into<String>`
+ - **Breaking**: `ArrayBuilder::dimension_names()` generalised to accept `Option<I>` where `I: IntoIterator<Item = D>` and `D: Into<DimensionName>`
+   - Can now write
+`builder.dimension_names(["y", "x"].into())` instead of `builder.dimension_names(vec!["y".into(), "x".into()].into())` 
 
 ### Removed
  - **Breaking**: remove `InvalidArraySubsetError` and `ArrayExtractElementsError`